fix: address CI failure and critical CodeRabbit security/correctness issues
- ci: switch npm ci → npm install to fix lockfile-out-of-sync failure
- security: enforce JWT_SECRET ≥ 32 chars at startup
- security: scope CORS Allow-Origin to configured domain instead of *
- security: HTML-escape user values in all email templates
- security: strip operator lat/lng from public API responses
- security: validate booking_url protocol (https only)
- correctness: fix || null → ?? null for falsy-zero fields (lat/lng/budget)
- correctness: fix nested sql template in Haversine query (Neon HTTP driver)
- correctness: block operators from updating client jobs (auth bypass)
- correctness: atomic UPDATE-as-gate in createBooking prevents race condition
- correctness: validate operator is verified before creating booking
- correctness: verify operator_profiles row exists before blob upload
- correctness: reset verification_status when FAA cert fields change
- correctness: save cert number/expiry before blob upload in onboarding
- correctness: cast COUNT/SUM to int in admin dashboard stats
- correctness: null-guard .container before appendChild in config.js
- email: prefer DEPLOY_PRIME_URL over URL for branch previews

+++ netlify/functions/api-admin.mjs
@@ -25,14 +25,14 @@ export default async (req, context) => {
 async function dashboard() {
   const [stats] = await sql`
     SELECT
-      (SELECT COUNT(*) FROM users WHERE role = 'client')   AS client_count,
-      (SELECT COUNT(*) FROM users WHERE role = 'operator') AS operator_count,
-      (SELECT COUNT(*) FROM operator_profiles WHERE verification_status = 'verified')      AS verified_operators,
-      (SELECT COUNT(*) FROM operator_profiles WHERE verification_status = 'under_review')  AS pending_verifications,
-      (SELECT COUNT(*) FROM jobs WHERE status = 'open')      AS open_jobs,
-      (SELECT COUNT(*) FROM jobs WHERE status = 'completed') AS completed_jobs,
-      (SELECT COUNT(*) FROM bookings WHERE status = 'completed') AS completed_bookings,
-      (SELECT COALESCE(SUM(platform_fee_cents), 0) FROM bookings WHERE status = 'completed') AS total_revenue_cents
+      (SELECT COUNT(*)::int FROM users WHERE role = 'client')   AS client_count,
+      (SELECT COUNT(*)::int FROM users WHERE role = 'operator') AS operator_count,
+      (SELECT COUNT(*)::int FROM operator_profiles WHERE verification_status = 'verified')      AS verified_operators,
+      (SELECT COUNT(*)::int FROM operator_profiles WHERE verification_status = 'under_review')  AS pending_verifications,
+      (SELECT COUNT(*)::int FROM jobs WHERE status = 'open')      AS open_jobs,
+      (SELECT COUNT(*)::int FROM jobs WHERE status = 'completed') AS completed_jobs,
+      (SELECT COUNT(*)::int FROM bookings WHERE status = 'completed') AS completed_bookings,
+      (SELECT COALESCE(SUM(platform_fee_cents), 0)::bigint FROM bookings WHERE status = 'completed') AS total_revenue_cents
   `
   return json(stats)
 }
